requirements.txt
Django~=2.2.25
PyMySQL~=1.1.0
bcrypt~=4.0.1
openpyxl~=3.1.2
pytz~=2023.3
djangorestframework~=3.9.1
pandas~=1.3.5
jieba~=0.42.1
pyecharts~=2.0.3
scikit-learn~=1.0.2
requests~=2.18.4
bs4~=0.0.1
beautifulsoup4~=4.9.3
djangorestframework-simplejwt~=5.2.2
django-cors-headers~=3.4.0